api/pyproject.toml
[tool.poetry]
name = "deploy history"
version = "0.1.0"
description = ""
authors = ["js.bae <js.bae@heumlabs.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "^0.20.0"
pygithub = "^1.57"
jira = "^3.4.1"
django-extensions = "^3.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-mock = "^3.10.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
